Functional hand model
Adding XrHandModel, an animated representation of the user's hands. It
lives within the XrOrigin coordinate system. This implementation
depends on various Meta-specific extensions, and is only tested on
the Quest 3.

Removes the handGeometry property, since that is useless without a
Skin, and it's non-trivial to make a declarative API for that.

Change-Id: I7263f042712cf15dcbdd0956d381775003a1c242
Reviewed-by: Christian Strømme <christian.stromme@qt.io>