Fix input processing for 2D Items in 3D scenes (internalPick)
Previously we would filter input events by using the childrenRect of the
Item2D and testing to see if the input event was inside of this rect.
This works fine for simple cases, but breaks down when the actual
contained items have any transforms applied to them, or the item uses
custom masking. There is already a system in place in Qt Quick to for
this so this patch uses that.

So first, the UV cordinate gives a pixel coordinate in the coordinate
space of the container item of the Item2D, so it's important to map that
pixel coordinate to the same coordinate space of the items inside of the
Item2D. Then by using the "contains" method of the QQuickItem under
test, this will work for the standard case (boundingRect) as well as the
masked case where the user creates a custom contains method themselves.

Pick-to: 6.5
Change-Id: I80fda5dea668d25e96560bee6f65d7238ae087fd
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit d63f1393b3e463c8b705609dc8fa659c529364ee)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>