Fix DeliveryAgent transform when View3D doesn't fill the whole window
There seems to be confusion with regard to coordinate spaces with the
DeliveryAgent transform code.  The DeliveryAgent is trying to map a
pointer event in scene space, but the code assumes that any View3D will
also be the same size as the window.  We must map the 2d scene space
coordinate first the View3D's space before setting up a raycast into
that
scene. This means that we *must* also provide a pointer to the View3D to
do that level of mapping. This should be fine considering that it is the
same View3D that creates the DeliveryAgent mapping anyway.

Without this fix, any input to a 3D scene that required grabbing would
be
very broken if the View3D wasn't the same size as the Window it was
inside of.

Pick-to: 6.5
Change-Id: I989deb0d9c857a9430f96039aaae454651efeb7e
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 316581cc44f3243fe61659df8883d1c744b971e1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>