Fix problem with using non-visible View3Ds as a sourceItem
Don't use the View3D's texture provider when the View3D is a hidden
sourceItem. Due to how the texture provider is set-up at the moment it's
driven by preprocess(), which is only called when the item is visible and
has a opacity higher then the cutoff threshold. To workaround this
limitation we now default to the path where we use a QSGLayer as an
intermediate step.

Fixes: QTBUG-122052
Change-Id: Idd6350b0c8475a55e84a3ec2badf180e2881c407
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>