Fix single pick methods when picking in scenes with 2d items
All of the pick methods that were expected to only return a single pick
result could be wrong when used in a scene with embedded 2d items. This
would happen when the nearest intersecting item was the plane of an
embedded 2d item. We would only return a single item to the frontend
that could be invalid (the hit was on the plane of a 2d item, but not
within its bounds). So it is important that all of our internal picking
methods always return a possible list of intersections for the case that
there are embedded 2d items.

This change makes it where the backend returns lists of possible hits,
and the frontend will return the best option.

Change-Id: I9f7984fee0fc1ee5fecc9c29b18082925ae8a0db
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>