Make correct use of the uvPosition property when picking Items
When picking 2D items embedded in the scene, the pickResult.uvPosition
property will contain the position value in the picked items coordinate
space. It makes sense to reuse the property for this because (we do it
already internally), and embedded items don't have a texture that
requires a UV, but the idea is the same.

This change makes sure that pickResult has the correct value mapped, and
this is now documented in pickResult.

Change-Id: I039c6890b008d6112f50d2c1858c87cd67d02b84
Reviewed-by: Christian Strømme <christian.stromme@qt.io>