Correction to index buffer type in loadMeshBVH
Only unsigned integer types are supported for user-supplied meshes in QQuick3DGeometry, and make sense for index buffers in general. The pre-set signed indexBufferFormat was always used in the old code, since "attribute.componentType" (either uint32 or uint16) never matched any of the signed alternatives in the if-clauses. This appears to have worked fine with uint32 buffers, but caused crashes with uint16 (due to out-of-bounds accesses caused by bad casts downstream).