Add support for submitting the depth buffer to the compositor
To achieve "submitting the depth buffer", the renderer has to follow
the same "pull" model as with color buffers (contrary to what
"submitting" would suggest): a separate XrSwapchain has to be created
for depth (depth-stencil), and then its images have to be acquired the
same way as with color images. Those are then the native textures (2D
texture, or texture array with multiview) to use as depth
(depth-stencil) buffer. (and data needs to be written out from the
render pass, it cannot just be potentially optimized away as we
normally do for depth-stencil)

The depth swapchain is always created whenever the extension is
supported. But its actual usage depends on the submit-depth flag,
which is a user controlled value, in form of a property on the XrView,
and can be changed dynamically at run time at any time.

When not enabled, no depth images are acquired and rendering uses its
own depth renderbuffer/textures as before. Only when enabled, do we
switch over to use the OpenXR-provided textures.  This may have
consequences for performance. The tradeoff between the better
reprojections enabled by submitting depth data and the potential
performance implications is not clear yet, but the expectation is that
there is a certain cost. (e.g. that depth data needs to be written out
even when the tiled GPU architecture could optimizate that way, and
with MSAA now depth data needs to be resolved too)

The property is false by default. Partly due to the perf. concerns and
also due to limited availability with certain configurations (MSAA,
see below). The env.var. QT_QUICK3D_XR_SUBMIT_DEPTH can be used to
make it true by default. Otherwise, it is up to the application to
decide to enable depth submission via the XrView property.

With MSAA enabled, resolving the depth-stencil multisample content
into a non-multisample depth texture (the one provided by OpenXR)
becomes mandatory. This cannot be supported with some APIs, such as
D3D, and Vulkan 1.0. On the Quest 3 (Android) this is not a problem
because it has Vulkan 1.1 + the relevant KHR extensions (but not the
Vulkan 1.2 core functionality without the KHR suffix), while OpenGL ES
should support this as well. When not supported, enabling depth
submission will be gracefully disabled, while also printing a
warning.

The XrSwapchain format selection is fixed up. With depth the order is
of higher importance, because we must prioritize choosing D24S8 over
other formats such as D16. Fix also the color format selection.

Task-number: QTBUG-122292
Change-Id: I85147b6fa1f9275f59e5b20d8d923310feb8d739
Reviewed-by: Andy Nichols <andy.nichols@qt.io>