Make sure we handle dirty texture in the effect
Texture changes were not handled in the effect, so only the initial
texture would be used. This problem only affects the Effect type as the
custom material already handles this. Note that unlike the custom
material this change takes a brute force approach and triggering a
"full-rebuild" of the effect, as that is the smallest reasonable change
that should fix this issue. This issue was discovered in DS when using the
ExtendedSceneEnvironment.

Pick-to: 6.7 6.6
Fixes: QTBUG-123442
Change-Id: I51310cd5dc5a6efd7681ee1ac035243a9b0dbd33
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Inho Lee <inho.lee@qt.io>