Enable usage of the IBL probe exposure with the ExtendedSceneEnvironment
Since the ExtendedSceneEnvironment is not using the built-in tonemapping of the
engine but instead uses its own custom tonemapping code, the internal tonemapping,
and therefore the exposure, was disabled. With this change the exposure and tonemapping
is decoupled when used with custom tonemapping code (internal only),
which allows the ExtendedSceneEnvironment to force the exposure values
to be added to the output.

Fixes: QTBUG-123846
Change-Id: I00cbdd85f4e3d3d58afff7899a5ca2ce61365a55
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>