Don't allow movement of ball after end of game in Quick Ball example
In the Quick Ball example, if you held onto the ball while the game
timer ran out, you would still be able to move the ball, throw it,
and even score more points that would update the point score that
had already been displayed.

After the game has finished, return the ball to its initial position.
Also, in the functions that control the movement of the ball, return
early if the game has finished, thus cancelling any operation done on
the ball if the user has not yet let go of the mouse-button.

Fixes: QTBUG-120001
Pick-to: 6.8 6.7
Change-Id: I23d63d718a682d430883c451be0c1efd23ca1716
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>