Do not reuse skybox passes for screen texture
The render passes for skybox/cubemap-based skybox
are not reusable always. A different sample count
for example (MSAA for the main target, no MSAA for
the screen texture) breaks everything since the
screen texture pass's skybox and the main target's
skybox are then need to be done with different
graphics pipeline settings. For simplicity, make
the screen texture pass to have its own skybox
passes instead of attempting to reuse the existing
ones from the main pass.

Add a Lancelot test as well.

Fixes: QTBUG-122890
Pick-to: 6.7
Change-Id: Ie44d6257596105837eb9490429d25bd4cc9ecbbd
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>