XrItem: press/release touch point based on vertical movement
Since the hand model will typically be locked to the plane, the user
will normally not know how far to move the hand to make the physical
tip of the finger leave the plane. Therefore, use relative vertical
movement to generate press/release events when the position is below the
plane.

Change-Id: I602aa05738288dd0bb16082bf16ce1f0d898e9cb
Reviewed-by: Andy Nichols <andy.nichols@qt.io>