Improve multiview support for depth pre-pass and depth texture
The Z prepass was working correctly already. Nonetheless, for
consistency we unify the helper prepare function arguments with
rhiPrepareRenderable by passing in the viewCount as well after samples.

The depth texture, for custom materials, is now exposed in two versions,
depending on multiview rendering being active. What is mildly annoying
that due to some of the underlying APIs we cannot make both work at
the same time, so user code that wishes to be compatible with multiview
has to be written with an appropriate ifdef:

 #if QSHADER_VIEW_COUNT >= 2
     vec4 depthSample = texture(DEPTH_TEXTURE_ARRAY, vec3(uv, VIEW_INDEX));
 #else
    vec4 depthSample = texture(DEPTH_TEXTURE, uv);
 #endif

Add a Lancelot test, although that cannot exercise the multiview path
with qt_depthTextureArray, obviously.

Add manual test assets for qmlxr then can be used to test the full
thing in action.

Update the docs somewhat. To be expanded later, because it is expected
that the screen texture needs a similar treatment.

Fixes: QTBUG-124843
Change-Id: I752c0ba2ce81c3e65175d9ea41816d288c368068
Reviewed-by: Andy Nichols <andy.nichols@qt.io>