Add internal machinery to set the local transform directly on a node
There are cases where we'll need to change the transform of a node
directly based on transforms coming from external APIs, so add some
internal helpers to aid with that. This is internal only, as the
approach used for e.g., decomposing the transform is fairly naive, which
we can live with as the only real reason we need to decompose the
transform is because we need to notify user if either scale, rotation or
position changes.

Change-Id: Ia81fc5f40212e199f986f5d2042d205bb4f40b4b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>