Improve the custom material multiview strategy
Do not flood the user (and our implementation, when it comes to enums
and such) with the _ARRAY variants. The original idea, where
DEPTH_TEXTURE would remain a sampler2D even in multiview mode, and would
still be functional with the left eye texture, is not feasible to do for
all the 3D APIs. And with that out of the picture, we can just let
DEPTH_TEXTURE expand to qt_depthTexture or qt_depthTextureArray
depending on having multiview mode enabled.

The details do not change, the user still needs to be aware of the type
(sampler2D vs sampler2DArray) and write code such as. However, the name
does not need to be varied now (DEPTH_TEXTURE in both cases)

  #if QSHADER_VIEW_COUNT >= 2
    vec4 depthSample = texture(DEPTH_TEXTURE, vec3(uv, VIEW_INDEX));
  #else
    vec4 depthSample = texture(DEPTH_TEXTURE, uv);
  #endif

Applies to SCREEN_TEXTURE and AO_TEXTURE as well, and, in the future, to
INPUT (qt_inputTexture[Array]) for postproc. effects.

Change-Id: I768fe9d48730f2d59023f86647d33365a52e53ae
Reviewed-by: Andy Nichols <andy.nichols@qt.io>