Enable postprocessing effects function in multiview mode
Also enables supersampling (SSAA) in multiview mode.
(not that anybody should use that in VR given the already high
resolution and GPU workload)

Using MSAA in combination with effects should work too.

Temporal and progressive AA are not covered in this patch.

No built-in effects are covered here. To make portable snippets,
once again one has to resort to ifdefing since INPUT is either
a 2D texture or a 2D texture array. Before this is added to the
built-in effects of ExtendedSceneEnvironment and elsewhere, those
will not function in multiview mode.

Further verification, with multi-pass effects in particular is also
going to happen separately. Basic effects, and combining them,
as demonstrated with the spiral + red colorize test for qmlxr, should
be functional.

Task-number: QTBUG-125056
Fixes: QTBUG-125132
Change-Id: I0640edde974d50863dd464b479b13cac30a8ee71
Reviewed-by: Christian Strømme <christian.stromme@qt.io>