Process also lines with macros and directives
Consider a custom material or effect snippet provided by the
application that works with some defines, for example:

  #define SAMPLE_LOD(uv, lod) textureLod(INPUT, vec3(uv, VIEW_INDEX), lod)

Here the substitution of INPUT and VIEW_INDEX does not currently happen
as expected. This is because the whole line, when starting with #, is
copied verbatim into the result, without applying the
identifier-by-identifier check that leads to replacing "VIEW_INDEX"
"qt_viewIndex" etc.

Change-Id: Ib41d0b06607cfa0ad6f8e927a1dee179d7048ab3
Reviewed-by: Christian Strømme <christian.stromme@qt.io>