Add support for flat varyings
It makes sense to enable this now, because the basic plumbing for
passing a flat uint qt_viewIndex from the vertex to the fragment
stage was implemented as part of the multiview work.

Extend this to user-specified input/output variables in the custom
material and effect shader snippets. One can now write
VARYING flat <type> <name>
i.e., prefix the type with "flat".

[ChangeLog][Quick3D] Custom materials and postprocessing effects
can now support passing data without interpolation between the vertex
and fragment stages. This is achieved by prefixing the type with the
'flat' keyword in the VARYING declaration.

Fixes: QTBUG-115972
Change-Id: Ifc07a2720998c865625ba540e30f745a45de2bb0
Reviewed-by: Christian Strømme <christian.stromme@qt.io>