Fix shadowed models despite recievesShadows=false
Even though Model::recievesShadows is set to false, sometimes the model
is rendered as completely shadowed. In the shadow mapping shader, the
light matrix is all zeroes when recievesShadows is false. As a
consequence of this we get a NaN value in 'smpCoord' and a division by
zero. We can mitigate this by simply checking if the 'shadowFactor' is
close to zero and just skip calling the shadow mapping function.

Pick-to: 6.5
Change-Id: I9ee6fe0219a7c4ca4b20c9cd24eaf9044d3d5fe6
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit c17345351f4ebc9a45d0eb1e3f70b0012297a0e8)
Reviewed-by: Jonas Karlsson <jonas.karlsson@qt.io>