Add multiview controlling property
The env.var stays and is documented as a shortcut to setting the
property to true. (and so qmlxr -w continues to work too)

The RuntimeInfo's is-multiview-really-in-use query is dropped in favor
of the XrView's enableMultiViewRendering since the value becomes true
also when using the env.var, so it can serve that purpose as well.
(regardless of how multiview mode is requested, the property value only
becomes true when multiview rendering is supported)

Besides the property docs, add also a dedicated page for multiview.

Like passthrough, VisionOS support is all dummy for now, and querying
always reports multiview to be unsupported.

Correct also some minor inconsistencies between the passthrough, depth
submission, and multiview getters and setters; these should be all
consistent with each other when it comes to naming and semantics.

Fixes: QTBUG-125600
Fixes: QTBUG-125599
Change-Id: I518f5b979fc3f33e029bd812534e87cbb60ce57c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>