Do not break with v65+ SDK
It is not enough to say that one can use the bundled OpenXR stuff, we
also need logic to ignore OCULUS_OPENXR_MOBILE_SDK when the loader .so
is not present (because it's a >= v65 SDK), and automatically use the
bundled loader also in this case.

Otherwise we end up configuring with non-bundled XR (even with SDK >=
v65) and then fail the build right away with a no rule to make the
(properietary) loader since the file is missing.

Change-Id: Iababd1f321ae0c4a99ed9c9e016d45b27ac11a3e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Elias Hautala <Elias.Hautala@qt.io>