Implement Cascading Shadow Maps (CSM)
Adds CSM support for directional lights. We use the Parallel Split
Shadow Maps (PSSM) technique for calculating the splits and limit the
number of cascades to four.

DirectionalLight adds the following properties for tweaking the CSM:
'csmSplit1', 'csmSplit2', 'csmSplit3', 'csmNumSplits' and
'csmBlendRatio'.

We now always add the eye view matrix to the fragment shaders since this
is potentially needed for calculating the cascade index for the
directional light.

A new algorithm for optimizing the shadow map size by intersecting the
view frustum and the casting and receiving scene bounding boxes.

The cones and cylinders test is modified for testing CSM.

Adds a lancelot test.

[ChangeLog] Add support for Cascading Shadow Maps in DirectionalLight.

Task-number: QTBUG-116885
Fixes: QTBUG-82333
Change-Id: I812c7d2de37f9dc12082e91f7736e05218d3807c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>