Handle OpenXR error results better
For function resolving, wrap the xrGetInstanceProcAddr into
a helper function that gives output such as this:

  OpenXR call failed (-7): XR_ERROR_FUNCTION_UNSUPPORTED
  Failed to resolve OpenXR function xrRequestSceneCaptureFB

Previously we only got the result string printed and nothing
more, which is insufficient when troubleshooting based on logs.

Similarly, when checking results from all other OpenXR calls, we
now get output such as this:

  OpenXR call failed (-7): XR_ERROR_FUNCTION_UNSUPPORTED
  Failed to get hand mesh data.

i.e. there is an indication from a handcrafted warning what
exactly goes wrong. The helper function is also marked as
nodiscard now.

Bailing out is handled on a case-by-case basis. Most failures
lead to returning from the function (perhaps with false or an
empty result), but for some this makes no sense and we can just
go on. This does not guarantee the application survives with
any kind of failures (that would need a deeper going through
of everything to make sure a failure to query, retrieve, create,
etc. something is handled gracefully), but at least the debug
output now has clear indication of which call went wrong.

It should be noted that some "errors" are not fatal, or
important even, e.g. when querying some interaction profile
bindings. This may be tuned later to make the output contain
fewer of these, perhaps misleading, "errors". (this is
relevant mainly with the XR Simulator)

Change-Id: Id406c2fee0a4f4be6c2783371230aa79d3f30b40
Reviewed-by: Andy Nichols <andy.nichols@qt.io>