Do not sort box in shadowmapping slicer
Let QSSGBounds3::toQSSGBoxPoints() return the points in the order that
the shadowmapping slicer expects. This way we avoid sorting it which can
cause the mesh to be invalid and in turn lead to invalid shadowmap
bounds.

Adding a lancelot test that showed the issue.

Amends: 0c85d183bd181c51ed7cd4ef1b3a5a3cc8543623

Change-Id: Ic8d4a332cd142adc78b16c423da25dbbc698cbe1
Reviewed-by: Andy Nichols <andy.nichols@qt.io>