Fix non-functional pickResult HitType enum
Previously because pickResult is a value type that defined an enum, the
expected behavior was that to use that enum you would do something like:
pickResult.Item
But that doesn't work because "pickResult.Item" wasn't defined and also
doesn't look quite right. So we follow the same pattern as other value
types with enums and use an uppercase version of the value type for the
enum values using a namespace. So now the same enum value would be:
PickResult.Item

Pick-to: 6.8
Change-Id: Ia3705db2ceb1eacdd16cfaaee76afa30b473aa5a
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>