Make sure XrItem is not pickable when not visible
For 3D models, visible and pickable are independent properties. A 2D
item in a Node is pickable if it is visible, even if the node is not
visible. XrItem has an internal Rectangle that is the parent of the
contentItem. This change ties that item's visibility to the visibility
of the XrItem.

Change-Id: I546e927ef68f8af525e85f4c935504c4ea0f7338
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
(cherry picked from commit 208d0d98966994b9c4cb009a1efab6063e0b37bf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>