Anchoring: Separate the platform code into separate implementations
The non-platform specific classes got renamed and the OpenXR named was
dropped as that's no longer accurate. File names were also changed to
reflect the new class names, but due to a unresolved issue with moc
qopenxrspatialanchor classes was not renamed this round; the name of the
file makes no difference in practice, it's just inconsistent.

Pick-to: 6.8
Change-Id: Ia06bd08f547c9cc78c6df374c2bd1514347e20d1
Reviewed-by: Christian Strømme <christian.stromme@qt.io>