Fix incorrect instancing model shadows
Previously the bounds for the scene did not take into account if a model
was used for instancing. To fix this we add a variable
'globalBoundsInstancing' to calculate the bounds of all the instances,
if present. We then use these bounds when calculating the bounds of the
shadow casting/receiving bounds of the model.

Pick-to: 6.8 6.7 6.5
Fixes: QTBUG-125876
Change-Id: I5d8917c289c1a296c908d1011306a0b83aff8f81
Reviewed-by: Andy Nichols <andy.nichols@qt.io>