Fix incorrect instancing model shadows
Previously the bounds for the scene did not take into account if a model
was used for instancing. To fix this we add a variable
'globalBoundsInstancing' to calculate the bounds of all the instances,
if present. We then use these bounds when calculating the bounds of the
shadow casting/receiving bounds of the model.

Pick-to: 6.5
Fixes: QTBUG-125876
Change-Id: I5d8917c289c1a296c908d1011306a0b83aff8f81
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 778e505fbb57231f5943716c906442240ce27b30)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit fc8ec8564bd8eebd92ccf07b0f38a6e2725d3d3c)