API: mark XrView::frameReady as internal
... and remove the QRhiTexture argument. The signal will be exposed even
if we don't document it. Having the QRhiTexture as an argument is also
not usable and has implications for how that is exposed and how it can
be used. With this change we keep the signal but resolve the texture
in place that one place we do use it (qmlxr).

Change-Id: Ide2128aba179ff98c1075c20714e4fe990e45938
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
(cherry picked from commit 0b85c051f6fde4d7198a75ae05b6567fdce2812c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>