Fix crash on exit with Models with morph targets and multiple subsets
If a model with multiple subsets contains morph targets, we stored a
pointer to the morph texture in each subset, but that texture is shared
between all subsets.  However on cleanup we tried to delete this texture
once per subset, which will lead to trying to delete the same texture
twice when a Model has more than one subset. This change makes sure that
we only delete the texture once.

Pick-to: 6.8 6.7 6.5
Change-Id: Ib170fe0bb67fc2ffb191ca7ed30f887cf8dd86ea
Reviewed-by: Christian Strømme <christian.stromme@qt.io>