API: Fix semantic labels
The semanticLables property is now renamed Classification and split
into two properties, where the common types are enumerated for better
cross platform usage while the string variant is left for types that
does not map, or that just won't be available in one platform over the
other. This way there's still a way to do classification of anchors
application side.

Pick-to: 6.8
Change-Id: Ic76d20a061d7cea74934aa84889f6c22ff2ea9e5
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>