Bind the boneTexture and/or morphTexture in shadow and depth passes
Bone/Skeleton and Morph target data is stored in textures, and since
these textures have an affect on the geometry of Models, it is important
that these textures be bound when setting up the Depth pre-pass and
Shadow passes. Previously the shaders tried to use these textures but no
data was bound, so we would get shader validation errors as well as
rendering errors.

Pick-to: 6.8
Fixes: QTBUG-126440
Change-Id: Id478535bd9408c04b8ebd877ea8cae11a1e40941
Reviewed-by: Inho Lee <inho.lee@qt.io>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>