Fix mismatched connect/disconnect styles issues
It seems it is quite important that you don't mix using the newer style
signal/slot connections, with the older string based style, as it can
lead to memory leaks. We were mixing the styles in many of our list
property functions, so that has been rectified, as well as fixing an
incorrect disconnection along the way.

Task-number: QTBUG-126581
Pick-to: 6.8 6.7 6.5
Change-Id: I238c6a4ff1709e38d826ad14afaf5c599482ea93
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>