Make sure to remove SRBs from cache when draw call data bindings change
Previously the QRhiShaderResourceBindings cache that we maintain could
end up growing indefinitely when Models were toggled between visible and
not. This occurred because when a Model is hidden, and we render some
frames without it, the resources associated with that model are released
(geometry meshes, textures, etc). When the model becomes visible again,
we would re-upload the necessary resources, but since Shader Resource
Bindings are cached with a key based on the pointer of the RHI resources
rather than the sources of those resources themselves, a new key for the
SRB would be generated, leaving the old SRB orphaned in the cache. Now
when this case occurs we make sure to remove the old SRB from the cache
before inserting the new one.

Task-number: QTBUG-126581
Pick-to: 6.5
Change-Id: Ifc02092c77f5e7168a2e3202d41aacbb24205952
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit f8f08aebdffb4fb8f62c30c0247d0dea23999143)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d2b530ed94ec4729770c00bb8870512374f7bde1)