materialeditor: support instancing
When using instanced model's customData, it can be used
only with CustomMaterial.
It will not fully support instancing's all the feature
but customData and color of InstanceEntryList are
enough to check the shading effect for current material.
They will be mapped to INSTANCE_DATA and INSTANCE_COLOR
in vertex shader.

When using instacing, Preview will show 5 instances and
the center object in position (0, 0, 0) is affected by
modified properties.(color and customData)
This object's index(INSTANCE_INDEX) is 0 and it can
be used in vertex shader, too.

Change-Id: Ied2fbf456366d52a41cb6ef2d1dbab266d38112a
Reviewed-by: Andy Nichols <andy.nichols@qt.io>