Fix the decompose function
The determinant can clearly not be two values at the same time... This
was causing the tracked camera to not emit position and rotation updates.
This function is quite simple as we have certain expectations of the data
we get, so just verify that the determinant isn't zero.

Change-Id: I8384548b01999f0fbf58582d8877e464f599339c
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
(cherry picked from commit 55bf1431107a0ce09457cebc67793d0e87244de3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>