VisionOS: Add code for pinch gestures
Mimic the gestures we already provide for OpenXR when it comes to
pinching (fingers touching the thumb).
Since only the index finger pinch is provided by VisionOS naively, we
need to detect this ourselves.

Known issues:
- We look for the gestures each frame as long as the hands
are tracked, regardless if the user has requested an action for a given
pinch gesture.
- We also detect the index finger pinch, even though this is already
  done by the system; the system event is processed differently.
- A modified version of the xr_simple example added as a manual test but
  set up as an example for easy testing on VisionOS at the moment.

Pick-to: 6.8
Change-Id: Ib7f98cc79c12648ea278165380c250a8d17c171b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>