Fix updateSpatialNode logic
updateSpatialNode() had some limitation and non-obvious issues when it
came to destroying or recreating the back-end objects. Depending on the
node type, and/or how a new object was created, it would either leak or
crash due to the internal mapping getting out of sync.
Internally this hasn't been a problem, as we don't actually re-create
back-end nodes anywhere; and user haven't really been able to create back-end
objects before the render extensions were added, but since that's now
exposed we need to make sure we handle it more gracefully.

Note that updateSpatialNode's behavior differs from QtQuick's
updatePaintNode's behavior with regards to destruction of back-end
objects, as QtQuick3D takes ownership of the object and destroys it
when: the front-end object is destroyed, or the back-end object changes.

Added tests to verify that we keep this working as described.

Change-Id: I1ef7344db229832315ca0bbd8466a898c74f501d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit ab875aca6ac1831a1f06ecaffa2e467c27fa529a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>