Better match up of the different hand enums
Ideally we should just have one, but that's not always convenient due
to header dependency or because we can't typedef enums and expose them
to QML. We should however strive to make it trivial to convert between
the ones we have. With this patch we match the different enums and also
move the somewhat special entry 'Unknown' to be after left and right,
as we are using the hand values as indices many places.

Change-Id: Ieec00ec481d965deda672d68e8d7935f2ef45a0b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
(cherry picked from commit f7e0f7ed6cbe49b64028561139abca6b27c1e82c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>