Doc: Fix warning for generating an empty group
There is no need to use the \generatelist command on a \group page to
list the members of the group; it happens automatically.

Remove the \generatelist command, and switch the order of sections to
accommodate any known issue topics that are added to the end of the
page.

Change-Id: I3c584a27d7c25e1035555816acd2945a1ecc7e14
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 8e7093a697e01cfa9e6a23ef393d231488edad78)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>