Remove the generationId from the texture data
The generationId is basically just a counter to check how "dirty" a
texture data instance is, and if it's dirty its rhi texture gets
released and replaced. Instead of having this separate dirty tracking we
now include the state in the hash and the size in the cache key, so any
incompatible changes will trigger a new entry to be added and a new
texture created. The old one will be released automatically later
if not used.

Pick-to: 6.7 6.5
Change-Id: I289a8ad135121b4313d90f073d5ee64a7bcbed51
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
Reviewed-by: Inho Lee <inho.lee@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit fd9e4834da57478dd31d387f568d40a3d22aa72f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>