Explicitly set the has-graphics-resource flag on all relevant types
This was the intention, but was left undone and forgotten, and the old
function in the base class isn't sufficient anymore, as we don't only
have distinct and internal types. The sub-classing of texture data
by the skin class demonstrates on of the problems of having this
predefined list of classes which gets tagged with the has-graphics-resource.

Pick-to: 6.8
Change-Id: Iab9e389287cb935e057c6349c548a37c4c22ee47
Reviewed-by: Andy Nichols <andy.nichols@qt.io>