Do not break with OpenXR 1.1 headers
The OpenXR SDK is OpenXR 1.1-aware from v36 (aptly versioned as 1.1.36,
as opposed to the previous 1.0.34). The bundled one in Qt is still
1.0.34.

So now building Qt Quick 3D against a newer version such as 1.1.40
succeeds, but there will be runtime problems, depending on the OpenXR
implementation used. Basically none of the examples will start with the
Meta XR Simulator, because xrCreateInstance fails always. Requesting API
1.1 on an 1.0 runtime is a bad idea.

For now this only affects those who build against non-bundled headers,
so one could say it is not critical, but the real problem is that once
the bundled headers are updated, things will break - and at runtime only
- and that can only be noticed by actually running applications against
all the relevant XR implementations (Simulator, Quest).

This is due to setting apiVersion to XR_CURRENT_API_VERSION. Which used
to be 1.0 with the bundled v34 headers, but is 1.1 with a newer set of
OpenXR headers. This will be rejected by an implementation that only
supports 1.0.

What exactly should be done here is unclear. For now, pass in 1.0(.34)
hardcoded. In the future, if OpenXR 1.1 features are needed (and
assuming runtimes will have hard checks for the apiVersion requested and
the features used), some more complicated solution will be needed.

It is worth noting that the the OpenXR versioning rules are somewhat
surprising. Including interesting deprecation and compatibility
promises, that aren't exactly matching Qt's.
https://registry.khronos.org/OpenXR/specs/1.0/html/xrspec.html#api-version-numbers-and-semantics

Change-Id: I01ed4f227b7518f3c4bd7153abe296d4a5955a3d
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 658c5911c7dbfcc9e22961806f33eceff220fbc5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>