Fix apalis-imx8 directional light shadow map issue
For some reason the apalis-imx8 reports that is supports R32F textures
and renders the shadowmap correctly but when sampling the texture in the
fragment shader it does not work. Until this is resolved  we revert to
using R16F as the highest texture shadow map precision.

Amends: b20fd4dcc707b7a18555d9b79193a9605d4a31c5

Pick-to: 6.8
Change-Id: I2c45c0281fcb83a4089de58726d6d892a3805ee8
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Kristoffer Skau <kristoffer.skau@qt.io>