Allow forcing mediump precision for floats in ES frag.shaders
Set the QT_QUICK3D_MEDIUM_PRECISION environment variable to a non-zero
value.

This is not meant for production use, but rather to allow quickly
measuring the results of medium precision (for floats) for the
entirety of the runtime generated fragment shaders on a given
device and GLES implementation.

[ChangeLog] Added the option of forcing medium precision for
floats in GLSL ES fragment shaders. Provided mainly as a tool
for performance tuning. Set the environment variable
QT_QUICK3D_MEDIUM_PRECISION to a non-zero value. Has no effect
on other kinds of shaders (and that includes non-ES GLSL too).

Change-Id: I9d3ecea3f84634dc15e94cea3043d80445781ded
Reviewed-by: Andy Nichols <andy.nichols@qt.io>