Fix frustum dimensions in shadowmap shader
The 'dimensionsInverted' variable used in the shadow mapping shader was
incorrectly calculated as half the size causing the pcfFactor and
shadowBias to act twice as big for DirectionalLight. The calculation is
now fixed which means that the pcfFactor and shadowBias is half the size
compared to previously. The default value of pcfFactor is doubled to
match the previous value but shadowBias is kept the same.

Fixes: QTBUG-128922
Change-Id: I73bc9e8edcaa73e04a4fd507ccf778a93ca26e32
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit bf3c208a6f57d72eab7b468296f8f206388d6ee7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>